css: new try for IE8 and image sizes
It seems that width+height fixed in `img` tag and `max-width` in CSS
will lead IE8 to squeeze images. The solution would be to add `width:
auto`. It seems to work for other browsers as well.

Also make it work in Firefox where max-width can only be applied to
block elements. Therefore, make the image `display: block` and also
add `max-width` to `p` (without and with Javascript).